refactor(libutil): convert ResultType to enum class
Convert ResultType from typedef enum to strongly-typed enum class:
- resFileLinked → ResultType::FileLinked
- resBuildLogLine → ResultType::BuildLogLine
- resUntrustedPath → ResultType::UntrustedPath
- resCorruptedPath → ResultType::CorruptedPath
- resSetPhase → ResultType::SetPhase
- resProgress → ResultType::Progress
- resSetExpected → ResultType::SetExpected
- resPostBuildLogLine → ResultType::PostBuildLogLine
- resFetchStatus → ResultType::FetchStatus

Add explicit cast in daemon.cc for wire protocol serialization.

+++ src/libstore/daemon.cc
@@ -174,7 +174,7 @@ struct TunnelLogger : public Logger
         if (GET_PROTOCOL_MINOR(clientVersion) < 20)
             return;
         StringSink buf;
-        buf << STDERR_RESULT << act << type << fields;
+        buf << STDERR_RESULT << act << static_cast<uint64_t>(type) << fields;
         enqueueMsg(buf.s);
     }
 };
